requirements.txt
numpy==1.26.4
pandas==2.2.2

ultralytics==8.3.85
tqdm==4.66.1
azure-storage-blob==12.19.0
pre-commit==4.1.0
exif==1.6.1
pyspark==3.5.5
pillow==11.1.0
